+++ Libplanet.Net.Tests/Consensus/Context/ContextTestBase.cs
@@ -83,7 +83,6 @@ namespace Libplanet.Net.Tests.Consensus.Context
             {
                 Index = BlockChain.Tip.Index + 1,
                 Difficulty = BlockChain.Tip.Difficulty,
-                TotalDifficulty = BlockChain.Tip.TotalDifficulty + BlockChain.Tip.Difficulty,
                 PublicKey = _fx.Miner.PublicKey,
                 PreviousHash = BlockChain.Tip.Hash,
                 Timestamp = BlockChain.Tip.Timestamp.Subtract(TimeSpan.FromSeconds(1)),