+++ Libplanet.Tests/TestUtils.cs
@@ -275,7 +275,6 @@ Actual (C# array lit):   new byte[{actual.LongLength}] {{ {actualRepr} }}";
             Assert.Equal(expected.Timestamp, actual.Timestamp);
             AssertBytesEqual(expected.Miner, actual.Miner);
             Assert.Equal(expected.Difficulty, actual.Difficulty);
-            Assert.Equal(expected.TotalDifficulty, actual.TotalDifficulty);
             AssertBytesEqual(expected.PreviousHash, actual.PreviousHash);
             AssertBytesEqual(expected.TxHash, actual.TxHash);
         }
@@ -412,7 +411,6 @@ Actual (C# array lit):   new byte[{actual.LongLength}] {{ {actualRepr} }}";
             {
                 Index = previousBlock.Index + 1,
                 Difficulty = difficulty,
-                TotalDifficulty = previousBlock.TotalDifficulty + difficulty,
                 Miner = miner?.ToAddress() ?? previousBlock.Miner,
                 PublicKey = protocolVersion < 2 ? null : miner ?? previousBlock.PublicKey,
                 PreviousHash = previousBlock.Hash,
@@ -444,7 +442,6 @@ Actual (C# array lit):   new byte[{actual.LongLength}] {{ {actualRepr} }}";
             {
                 Index = previousBlock.Index + 1,
                 Difficulty = 1, // 0 difficulty is not allowed.
-                TotalDifficulty = previousBlock.TotalDifficulty + 1,
                 Miner = miner?.ToAddress() ?? previousBlock.Miner,
                 PublicKey = protocolVersion < 2 ? null : miner ?? previousBlock.PublicKey,
                 PreviousHash = previousBlock.Hash,