+++ Libplanet/Blocks/PreEvaluationBlock.cs
@@ -45,9 +45,6 @@ namespace Libplanet.Blocks
         /// </exception>
         /// <exception cref="InvalidBlockDifficultyException">Thrown when the value to set is
         ///  is negative.</exception>
-        /// <exception cref="InvalidBlockTotalDifficultyException">Thrown when
-        /// the <paramref name="content"/>'s <see cref="IBlockMetadata.TotalDifficulty"/> is less
-        /// than its <see cref="IBlockMetadata.Difficulty"/>.</exception>
         /// <exception cref="InvalidTxSignatureException">Thrown when any tx signature is invalid or
         /// not signed by its signer.</exception>
         /// <exception cref="InvalidTxNonceException">Thrown when the same tx nonce is used by
@@ -93,9 +90,6 @@ namespace Libplanet.Blocks
         /// </exception>
         /// <exception cref="InvalidBlockDifficultyException">Thrown when the value to set is
         ///  is negative.</exception>
-        /// <exception cref="InvalidBlockTotalDifficultyException">Thrown when
-        /// the <paramref name="content"/>'s <see cref="IBlockMetadata.TotalDifficulty"/> is less
-        /// than its <see cref="IBlockMetadata.Difficulty"/>.</exception>
         /// <exception cref="InvalidTxSignatureException">Thrown when any tx signature is invalid or
         /// not signed by its signer.</exception>
         /// <exception cref="InvalidTxNonceException">Thrown when the same tx nonce is used by